requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pyjwt>=2.8.0
networkx>=3.1
nltk>=3.8.1
altair>=5.0.0
scipy>=1.10.0
